requirements.txt
torch
transformers
Pillow
torchvision
gradio
huggingface-hub
opencv-python
flask
dill
requests
#-e